box: add same_reference_as() & default constructor
Add a check (and an overloaded operator==) which compares the held
references to another box<T>. This means that comparing the box<T>s
returns true if both store the same object.

The box<T> now has a default constructor, which allows it to exist
in a list<T> and more coming types.